requirements.txt
python-binance
ccxt
python-dotenv
tabulate
schedule
Flask
Flask-Login
werkzeug
requests
gunicorn
# sqlite3 is part of the Python standard library
